chromeos-base/chromeos-base: handle /etc/localtime when USE=-cros_host
The /etc/localtime symlink is unconditionally installed.  While we want
this behavior for the boards, we don't want this for the cros sdk as we
manually setup that file from the user's external system.  So add the
cros_host USE flag to control the behavior.

BUG=chromium-os:21850
TEST=`emerge chromeos-base` no longer clobbers /etc/localtime

Reviewed-on: https://gerrit.chromium.org/gerrit/10960
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>

+++ chromeos-base/chromeos-base/chromeos-base-0.ebuild
@@ -10,7 +10,7 @@ SRC_URI=""
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="amd64 arm x86"
-IUSE=""
+IUSE="cros_host"
 
 DEPEND=">=sys-apps/baselayout-2"
 RDEPEND="${DEPEND}
@@ -72,10 +72,12 @@ add_users_to_group() {
 }
 
 pkg_setup() {
-	# The sys-libs/timezone-data package installs a default /etc/localtime
-	# file automatically, so scrub that if it's a regular file.
-	local etc_tz="${ROOT}etc/localtime"
-	[[ -L ${etc_tz} ]] || rm -f "${etc_tz}"
+	if ! use cros_host ; then
+		# The sys-libs/timezone-data package installs a default /etc/localtime
+		# file automatically, so scrub that if it's a regular file.
+		local etc_tz="${ROOT}etc/localtime"
+		[[ -L ${etc_tz} ]] || rm -f "${etc_tz}"
+	fi
 }
 
 src_install() {
@@ -85,9 +87,12 @@ src_install() {
 	insinto /etc/profile.d
 	doins "${FILESDIR}"/xauthority.sh || die
 
-	# Symlink /etc/localtime to something on the stateful partition, which we
-	# can then change around at runtime.
-	dosym /var/lib/timezone/localtime /etc/localtime
+	# target-specific fun
+	if ! use cros_host ; then
+		# Symlink /etc/localtime to something on the stateful partition, which we
+		# can then change around at runtime.
+		dosym /var/lib/timezone/localtime /etc/localtime || die
+	fi
 }
 
 pkg_postinst() {
